Moving utility method out of BudgetLineItemStatus enum to utility class to prevent interpreter error.
