+++ backend/models/cans.py
@@ -40,18 +40,6 @@ class BudgetLineItemStatus(Enum):
     PLANNED = "Planned"
     IN_EXECUTION = "In Execution"
     OBLIGATED = "Obligated"
-    @staticmethod
-    def convert_BLI_status_name_to_pretty_string(status_name):
-        if status_name == "DRAFT":
-            return BudgetLineItemStatus.DRAFT.__str__()
-        elif status_name == "PLANNED":
-            return BudgetLineItemStatus.PLANNED.__str__()
-        elif status_name == "IN_EXECUTION":
-            return BudgetLineItemStatus.IN_EXECUTION.__str__()
-        elif status_name == "OBLIGATED":
-            return BudgetLineItemStatus.OBLIGATED.__str__()
-        else:
-            return BudgetLineItemStatus.DRAFT.__str__()
 
 class ModType(Enum):
     ADMIN = auto()
