+++ backend/ops_api/ops/utils/budget_line_items.py
@@ -1,6 +1,6 @@
 from flask import current_app
 
-from models import CAN, BudgetLineItem, Division, Portfolio
+from models import CAN, BudgetLineItem, BudgetLineItemStatus, Division, Portfolio
 
 
 def get_division_for_budget_line_item(bli_id: int):
@@ -13,3 +13,16 @@ def get_division_for_budget_line_item(bli_id: int):
         .one_or_none()
     )
     return division
+
+
+def convert_BLI_status_name_to_pretty_string(status_name):
+    if status_name == "DRAFT":
+        return BudgetLineItemStatus.DRAFT.__str__()
+    elif status_name == "PLANNED":
+        return BudgetLineItemStatus.PLANNED.__str__()
+    elif status_name == "IN_EXECUTION":
+        return BudgetLineItemStatus.IN_EXECUTION.__str__()
+    elif status_name == "OBLIGATED":
+        return BudgetLineItemStatus.OBLIGATED.__str__()
+    else:
+        return BudgetLineItemStatus.DRAFT.__str__()
