+++ backend/ops_api/ops/utils/change_requests.py
@@ -11,6 +11,7 @@ from models import (
     Division,
 )
 from models.cans import BudgetLineItemStatus
+from ops_api.ops.utils.budget_line_items import convert_BLI_status_name_to_pretty_string
 
 
 def get_expires_date():
@@ -65,8 +66,8 @@ def create_notification_of_reviews_request_to_submitter(change_request: ChangeRe
 
     if change_request.has_status_change:
         status_diff = change_request.requested_change_diff["status"]
-        new_status = BudgetLineItemStatus.convert_BLI_status_name_to_pretty_string(status_diff["new"])
-        old_status = BudgetLineItemStatus.convert_BLI_status_name_to_pretty_string(status_diff["old"])
+        new_status = convert_BLI_status_name_to_pretty_string(status_diff["new"])
+        old_status = convert_BLI_status_name_to_pretty_string(status_diff["old"])
         if change_request.status == ChangeRequestStatus.APPROVED:
             notification = ChangeRequestNotification(
                 change_request_id=change_request.id,
